+++ src/test/java/ai/labs/eddi/integration/AgentDeploymentComponentIT.java
@@ -93,7 +93,7 @@ public class AgentDeploymentComponentIT extends BaseIntegrationIT {
 
         String packageBody = String.format("""
                 {
-                  "WorkflowSteps": [
+                  "workflowSteps": [
                     {
                       "type": "eddi://ai.labs.parser",
                       "config": {},