+++ src/test/java/ai/labs/eddi/integration/ApiContractIT.java
@@ -126,7 +126,7 @@ public class ApiContractIT extends BaseIntegrationIT {
 
         String packageBody = String.format("""
                 {
-                  "WorkflowSteps": [
+                  "workflowSteps": [
                     {
                       "type": "eddi://ai.labs.parser",
                       "config": {},