+++ src/test/java/ai/labs/eddi/integration/ConversationServiceComponentIT.java
@@ -153,7 +153,7 @@ public class ConversationServiceComponentIT extends BaseIntegrationIT {
 
         String packageBody = String.format("""
                 {
-                  "WorkflowSteps": [
+                  "workflowSteps": [
                     {
                       "type": "eddi://ai.labs.parser",
                       "config": {},